Move some validation code into submodules
