+++ cedar-policy-validator/src/typecheck.rs
@@ -18,16 +18,10 @@
 //! the `Typechecker` struct by calling the `typecheck_policy` method given a
 //! policy.
 
-mod test_expr;
-mod test_extensions;
-mod test_namespace;
-mod test_optional_attributes;
-mod test_partial;
-mod test_policy;
-mod test_strict;
-mod test_type_annotation;
-mod test_unspecified_entity;
-mod test_utils;
+mod test;
+
+mod typecheck_answer;
+pub(crate) use typecheck_answer::TypecheckAnswer;
 
 use std::{
     borrow::Cow,
@@ -54,176 +48,6 @@ use cedar_policy_core::ast::{
 
 #[cfg(not(target_arch = "wasm32"))]
 const REQUIRED_STACK_SPACE: usize = 1024 * 100;
-
-/// TypecheckAnswer holds the result of typechecking an expression.
-#[derive(Debug, Eq, PartialEq)]
-pub(crate) enum TypecheckAnswer<'a> {
-    /// Typechecking succeeded, and we know the type and a possibly empty effect
-    /// set for the expression. The effect set is the set of
-    /// (expression, attribute) pairs that are known as safe to access under the
-    /// assumption that the expression evaluates to true.
-    TypecheckSuccess {
-        expr_type: Expr<Option<Type>>,
-        expr_effect: EffectSet<'a>,
-    },
-    /// Typechecking failed. We might still be able to know the type of the
-    /// overall expression, but not always. For instance, an `&&` expression
-    /// will always have type `boolean`, so we populate `expr_recovery_type`
-    /// with `Some(boolean)` even when there is a type error in the expression.
-    TypecheckFail {
-        expr_recovery_type: Expr<Option<Type>>,
-    },
-
-    /// RecursionLimit Reached
-    RecursionLimit,
-}
-
-impl<'a> TypecheckAnswer<'a> {
-    /// Construct a successful TypecheckAnswer with a type but with an empty
-    /// effect set.
-    pub fn success(expr_type: Expr<Option<Type>>) -> Self {
-        Self::TypecheckSuccess {
-            expr_type,
-            expr_effect: EffectSet::new(),
-        }
-    }
-
-    /// Construct a successful TypecheckAnswer with a type and an effect.
-    pub fn success_with_effect(expr_type: Expr<Option<Type>>, expr_effect: EffectSet<'a>) -> Self {
-        Self::TypecheckSuccess {
-            expr_type,
-            expr_effect,
-        }
-    }
-
-    /// Construct a failing TypecheckAnswer with a type.
-    pub fn fail(expr_type: Expr<Option<Type>>) -> Self {
-        Self::TypecheckFail {
-            expr_recovery_type: expr_type,
-        }
-    }
-
-    /// Check if this TypecheckAnswer contains a particular type. It
-    /// contains a type if the type annotated AST contains `Some`
-    /// of the argument type at its root.
-    pub fn contains_type(&self, ty: &Type) -> bool {
-        match self {
-            TypecheckAnswer::TypecheckSuccess { expr_type, .. } => Some(expr_type),
-            TypecheckAnswer::TypecheckFail { expr_recovery_type } => Some(expr_recovery_type),
-            TypecheckAnswer::RecursionLimit => None,
-        }
-        .and_then(|e| e.data().as_ref())
-            == Some(ty)
-    }
-
-    pub fn into_typed_expr(self) -> Option<Expr<Option<Type>>> {
-        match self {
-            TypecheckAnswer::TypecheckSuccess { expr_type, .. } => Some(expr_type),
-            TypecheckAnswer::TypecheckFail { expr_recovery_type } => Some(expr_recovery_type),
-            TypecheckAnswer::RecursionLimit => None,
-        }
-    }
-
-    /// Return true if this represents successful typechecking.
-    pub fn typechecked(&self) -> bool {
-        match self {
-            TypecheckAnswer::TypecheckSuccess { .. } => true,
-            TypecheckAnswer::TypecheckFail { .. } => false,
-            TypecheckAnswer::RecursionLimit => false,
-        }
-    }
-
-    /// Transform the effect of this TypecheckAnswer without modifying the
-    /// success or type.
-    pub fn map_effect<F>(self, f: F) -> Self
-    where
-        F: FnOnce(EffectSet<'a>) -> EffectSet<'a>,
-    {
-        match self {
-            TypecheckAnswer::TypecheckSuccess {
-                expr_type,
-                expr_effect,
-            } => TypecheckAnswer::TypecheckSuccess {
-                expr_type,
-                expr_effect: f(expr_effect),
-            },
-            TypecheckAnswer::TypecheckFail { .. } => self,
-            TypecheckAnswer::RecursionLimit => self,
-        }
-    }
-
-    /// Convert this TypecheckAnswer into an equivalent answer for an expression
-    /// that has failed to typecheck. If this is already TypecheckFail, then no
-    /// change is required, otherwise, a TypecheckFail is constructed containing
-    /// `Some` of the `expr_type`.
-    pub fn into_fail(self) -> Self {
-        match self {
-            TypecheckAnswer::TypecheckSuccess { expr_type, .. } => TypecheckAnswer::fail(expr_type),
-            TypecheckAnswer::TypecheckFail { .. } => self,
-            TypecheckAnswer::RecursionLimit => self,
-        }
-    }
-
-    /// Sequence another typechecking operation after this answer. The result of
-    /// the operation will be adjusted to be a TypecheckFail if this is a
-    /// TypecheckFail, otherwise it will be returned unaltered.
-    pub fn then_typecheck<F>(self, f: F) -> Self
-    where
-        F: FnOnce(Expr<Option<Type>>, EffectSet<'a>) -> TypecheckAnswer<'a>,
-    {
-        match self {
-            TypecheckAnswer::TypecheckSuccess {
-                expr_type,
-                expr_effect,
-            } => f(expr_type, expr_effect),
-            TypecheckAnswer::TypecheckFail { expr_recovery_type } => {
-                f(expr_recovery_type, EffectSet::new()).into_fail()
-            }
-            TypecheckAnswer::RecursionLimit => self,
-        }
-    }
-
-    /// Sequence another typechecking operation after all of the typechecking
-    /// answers in the argument. The result of the operation is adjusted in the
-    /// same manner as in `then_typecheck`, but accounts for the all the
-    /// TypecheckAnswers.
-    pub fn sequence_all_then_typecheck<F>(
-        answers: impl IntoIterator<Item = TypecheckAnswer<'a>>,
-        f: F,
-    ) -> TypecheckAnswer<'a>
-    where
-        F: FnOnce(Vec<(Expr<Option<Type>>, EffectSet<'a>)>) -> TypecheckAnswer<'a>,
-    {
-        let mut unwrapped = Vec::new();
-        let mut any_failed = false;
-        let mut recusion_limit_reached = false;
-        for ans in answers {
-            any_failed |= !ans.typechecked();
-            unwrapped.push(match ans {
-                TypecheckAnswer::TypecheckSuccess {
-                    expr_type,
-                    expr_effect,
-                } => (expr_type, expr_effect),
-                TypecheckAnswer::TypecheckFail { expr_recovery_type } => {
-                    (expr_recovery_type, EffectSet::new())
-                }
-                TypecheckAnswer::RecursionLimit => {
-                    recusion_limit_reached = true;
-                    break;
-                }
-            });
-        }
-
-        let ans = f(unwrapped);
-        if recusion_limit_reached {
-            TypecheckAnswer::RecursionLimit
-        } else if any_failed {
-            ans.into_fail()
-        } else {
-            ans
-        }
-    }
-}
 
 /// Basic result for typechecking
 #[derive(Debug)]
